legacy/python-fastapi/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Bick Hsu", email = "bickhsu@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "google-generativeai>=0.8.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.44",
    "supabase>=2.28.3",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
